⚡️: fix Answer, Question 관련 오류 수정 #38
