+++ plu-domain/src/main/kotlin/com/th/plu/domain/domain/answer/repository/AnswerRepositoryImpl.kt
@@ -13,48 +13,48 @@ import org.springframework.stereotype.Repository
 class AnswerRepositoryImpl(private val queryFactory: JPAQueryFactory) : AnswerRepositoryCustom {
     override fun findAnswerById(id: Long): Answer? {
         return queryFactory
-                .selectFrom(answer)
-                .where(answer.id.eq(id))
-                .fetchOne()
+            .selectFrom(answer)
+            .where(answer._id.eq(id))
+            .fetchOne()
     }
 
     override fun findEveryAnswersWithCursorAndPageSize(questionId: Long, lastAnswerId: Long, pageSize: Long): List<EveryAnswerRetrieveResponse> {
         return queryFactory
-                .select(QEveryAnswerRetrieveResponse(answer.id, like.answer.id.count(), answer.content))
-                .from(answer)
-                .leftJoin(like).on(like.answer.id.eq(answer.id))
-                .where(
-                        answer.isPublic.eq(true),
-                        answer.question.id.eq(questionId),
-                        answer.id.lt(lastAnswerId),
-                )
-                .groupBy(answer.id)
-                .orderBy(answer.id.desc())
-                .limit(pageSize)
-                .fetch()
+            .select(QEveryAnswerRetrieveResponse(answer._id, like.answer._id.count(), answer.content))
+            .from(answer)
+            .leftJoin(like).on(like.answer._id.eq(answer._id))
+            .where(
+                answer.isPublic.eq(true),
+                answer.question._id.eq(questionId),
+                answer._id.lt(lastAnswerId),
+            )
+            .groupBy(answer._id)
+            .orderBy(answer._id.desc())
+            .limit(pageSize)
+            .fetch()
     }
 
     override fun findPublicAnswersCountByQuestionId(questionId: Long): Long {
         return queryFactory
-                .select(answer.id.count())
-                .from(answer)
-                .where(answer.question.id.eq(questionId))
-                .fetchOne()!!
+            .select(answer._id.count())
+            .from(answer)
+            .where(answer.question._id.eq(questionId))
+            .fetchOne()!!
     }
 
     override fun findPublicAnswersLikeTopN(questionId: Long, getCount: Long): List<EveryAnswerRetrieveResponse> {
         return queryFactory
-                .select(QEveryAnswerRetrieveResponse(answer.id, like.answer.id.count(), answer.content))
-                .from(answer)
-                .leftJoin(like).on(like.answer.id.eq(answer.id))
-                .where(
-                        answer.isPublic.eq(true),
-                        answer.question.id.eq(questionId)
-                )
-                .groupBy(answer.id)
-                .orderBy(like.answer.id.count().desc())
-                .limit(getCount)
-                .fetch()
+            .select(QEveryAnswerRetrieveResponse(answer._id, like.answer._id.count(), answer.content))
+            .from(answer)
+            .leftJoin(like).on(like.answer._id.eq(answer._id))
+            .where(
+                answer.isPublic.eq(true),
+                answer.question._id.eq(questionId)
+            )
+            .groupBy(answer._id)
+            .orderBy(like.answer._id.count().desc())
+            .limit(getCount)
+            .fetch()
     }
 
     override fun existsByMemberIdAndQuestionId(memberId: Long, questionId: Long): Boolean {
